criando usuario admin padrao para o sistema

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -54,7 +54,7 @@ public class DAOUsuarioRepository {
 
     public List<ModelLogin> consultaUsuarioList() throws SQLException {
         List<ModelLogin> retorno = new ArrayList<>();
-        String sql = "select * from model_login";
+        String sql = "select * from model_login where useradmin is false";
         PreparedStatement statement = connection.prepareStatement(sql);
         ResultSet resultado = statement.executeQuery();
 
@@ -73,7 +73,7 @@ public class DAOUsuarioRepository {
 
     public List<ModelLogin> consultaUsuarioList(String nome) throws SQLException {
     	List<ModelLogin> retorno = new ArrayList<ModelLogin>();
-    	String sql = "select * from model_login where upper(nome) like upper(?)";
+    	String sql = "select * from model_login where upper(nome) like upper(?) and useradmin is false";
     	PreparedStatement statement = connection.prepareStatement(sql);
     	statement.setString(1, "%"+nome+"%");
     	
@@ -95,7 +95,7 @@ public class DAOUsuarioRepository {
     public ModelLogin consultaUsuarioporId(String id) throws Exception {
         ModelLogin modelLogin = new ModelLogin();
 
-        String sql = "SELECT * FROM model_login WHERE id = ?";
+        String sql = "SELECT * FROM model_login WHERE id = ? and useradmin is false";
         PreparedStatement statement = connection.prepareStatement(sql);
         statement.setLong(1, Long.parseLong(id));
 
@@ -114,7 +114,7 @@ public class DAOUsuarioRepository {
     public ModelLogin consultaUsuario(String login) throws Exception {
         ModelLogin modelLogin = new ModelLogin();
 
-        String sql = "SELECT * FROM model_login WHERE login = ?";
+        String sql = "SELECT * FROM model_login WHERE login = ? and useradmin is false";
         PreparedStatement statement = connection.prepareStatement(sql);
         statement.setString(1, login);
 
@@ -140,7 +140,7 @@ public class DAOUsuarioRepository {
     }
 
     public void deletarUsuario(String idUser) throws Exception{
-        String sql = "DELETE FROM model_login WHERE id = ?;";
+        String sql = "DELETE FROM model_login WHERE id = ? and useradmin is false;";
         PreparedStatement preparedStatement = connection.prepareStatement(sql);
         preparedStatement.setLong( 1, Long.parseLong(idUser));
